requirements.txt
tensorflow==2.20.0
fastapi
uvicorn
pillow
mlflow
scikit-learn
matplotlib
seaborn
numpy
python-multipart